requirements.txt
selenium
python-dotenv
pywhatkit
pymongo
webdriver_manager
pyautogui